Avancement gestion des thèmes

+++ Xalise.Web/Enums/EnumModeOuverture.cs
@@ -5,7 +5,7 @@ namespace Xalise.Web.Enums
     /// <summary>
     /// Mode d'ouverture des pages et des fenêtres de dialogue.
     /// </summary>
-    public enum eModeOuverture : int
+    public enum eModeOuverture : short
     {
         [Description("Création")]
         CREATION        = 1,
